Create README.md
The goal of this project is to create a program that uses machine learning to uncover persons of interest (POI), given data on former Enron employees. A person of interest is defined as a person who might have been participating in fraudulent behavior when working at Enron. Because emails and financial records were made public on employees of Enron through legal investigations of widespread fraud, we have a lot of data to work with. Machine learning is appropriate for the task of finding persons of interest, because we can use the data we have on predefined persons of interest in order to teach an algorithm to identify new points of interest. For example, I found through my algorithm that total stock value, bonus, and exercised stock options data could be used as strong predictors for those that might have been involved in fraud, thus if I have financial data on all employees at Enron, I could use the algorithm to identify new persons of interest. 
The dataset I am using has 146 data points (i.e. people). There are 21 features available for each person. There are 18 predetermined persons of interest in the dataset (out of the 146).